+++ web/src/components/Buttons2/RemediationButton/graphql/remediateResource.generated.ts
