+++ web/src/components/forms/BaseRuleForm/BaseRuleForm.tsx
@@ -18,7 +18,7 @@
 
 import React from 'react';
 import { Formik } from 'formik';
-import { SubmitButton } from 'Components/Buttons';
+import SubmitButton from 'Components/Buttons2/SubmitButton';
 import * as Yup from 'yup';
 import { Flex, Button } from 'pouncejs';
 import useRouter from 'Hooks/useRouter';
