+++ web/src/components/forms/ChangePasswordForm/ChangePasswordForm.tsx
@@ -21,7 +21,7 @@ import { Field, Formik } from 'formik';
 import React from 'react';
 import * as Yup from 'yup';
 import { createYupPasswordValidationSchema } from 'Helpers/utils';
-import { SubmitButton } from 'Components/Buttons';
+import SubmitButton from 'Components/Buttons2/SubmitButton';
 import FormikTextInput from 'Components/fields/TextInput';
 import useAuth from 'Hooks/useAuth';
