+++ web/src/components/forms/SetPasswordForm/SetPasswordForm.tsx
@@ -21,7 +21,7 @@ import React from 'react';
 import * as Yup from 'yup';
 import { createYupPasswordValidationSchema } from 'Helpers/utils';
 import { Alert, Box, Text } from 'pouncejs';
-import { SubmitButton } from 'Components/Buttons';
+import SubmitButton from 'Components/Buttons2/SubmitButton';
 import FormikTextInput from 'Components/fields/TextInput';
 import useAuth from 'Hooks/useAuth';
