+++ web/src/pages/CreateComplianceSource/SuccessPanel/SuccessPanel.tsx
@@ -18,7 +18,7 @@
 
 import React from 'react';
 import { Flex, Heading, Text } from 'pouncejs';
-import { SubmitButton } from 'Components/Buttons';
+import SubmitButton from 'Components/Buttons2/SubmitButton';
 
 interface SuccessPanelProps {
   loading: boolean;
