+++ web/src/pages/ResourceDetails/columns.tsx
@@ -23,7 +23,8 @@ import { Text, TableProps, Badge, Tooltip, Label } from 'pouncejs';
 import { ComplianceItem, ComplianceStatusEnum } from 'Generated/schema';
 import { capitalize } from 'Helpers/utils';
 import { SEVERITY_COLOR_MAP } from 'Source/constants';
-import { SuppressButton, RemediationButton } from 'Components/Buttons';
+import SuppressButton from 'Components/Buttons2/SuppressButton';
+import RemediationButton from 'Components/Buttons2/RemediationButton';
 
 // The columns that the associated table will show
 const columns = [
